requirements.txt
discord.py==1.7.3
instagrapi==1.16.16
python-dotenv==0.20.0
